FIX: Disabled Swagger generation
This appears to be out of step with the Dropwizard version and is causing issues. For now it's commented out and disabled in POM and sources until we get the application up to date. Once the dependencies are updated and the CI workflow is ready we will look to get this working again.